+++ town.cpp
@@ -82,7 +82,7 @@ void TOWN::upgradeBuild()
 			if (_pPlayer->getTree() <= _pBuildMgr->getConsumptionResource(E_BUILDS::E_KEEP, E_RESOURCE::E_TREE)) return;
 			if (_pPlayer->getOil() <= _pBuildMgr->getConsumptionResource(E_BUILDS::E_KEEP, E_RESOURCE::E_OIL)) return;
 
-			_pBuildMgr->buildBuilding(E_BUILDS::E_KEEP, OBJECT::getLeft(), OBJECT::getTop());
+			_pBuildMgr->buildBuilding(E_BUILDS::E_KEEP, (float)OBJECT::getLeft(), (float)OBJECT::getTop());
 
 			_pBuildMgr->removeBuild(this);
 		}
